yolo/requirements.txt
ultralytics
boto3
pyyaml
mlflow
